Separa clases de la gui
