+++ gui/gui.py
@@ -1,6 +1,8 @@
-import subprocess
 import customtkinter as ctk
-from datetime import datetime
+from querier_page import QuerierPage
+from history_page import HistoryPage
+from log_frame import LogsFrame
+from indexer_page import IndexerPage
 
 ctk.set_appearance_mode("System")  # Modes: "System" (standard), "Dark", "Light"
 ctk.set_default_color_theme("blue")  # Themes: "blue" (standard), "green", "dark-blue"
@@ -134,190 +136,6 @@ class App(ctk.CTk):
         # FIXME: Esto es solo un botón de juguete para probar funcionalidades
         print("HISTORY BUTOOTN click")
         self.show_frame("HistoryPage")
-
-
-class QuerierPage(ctk.CTkFrame):
-    def __init__(self, parent, controller):
-        ctk.CTkFrame.__init__(self, parent)
-        self.controller = controller
-        # label = ctk.CTkLabel(self, text="This is the Querier Page")
-        # label.pack(side="top", fill="x", pady=10)
-
-        # button1 = ctk.CTkButton(self, text="Go to Page One",
-        #                     command=lambda: controller.show_frame("PageOne"))
-        # button2 = ctk.CTkButton(self, text="Go to Page Two",
-        #                     command=lambda: controller.show_frame("PageTwo"))
-        # button1.pack()
-        # button2.pack()
-
-        # Tab Querier
-        self.querier_tab_title_lbl = ctk.CTkLabel(self, text="Title", anchor="w",
-                                                  font=ctk.CTkFont(weight="bold"))
-        self.querier_tab_title_lbl.pack(side="top", anchor="w", padx=20, pady=0)
-        self.querier_tab_entry_title = ctk.CTkEntry(self, placeholder_text="Inserte texto aquí")
-        self.querier_tab_entry_title.pack(side="top", padx=20, pady=5, fill="x")
-
-        self.querier_tab_abs_lbl = ctk.CTkLabel(self, text="Abstract", anchor="w",
-                                                font=ctk.CTkFont(weight="bold"))
-        self.querier_tab_abs_lbl.pack(side="top", anchor="w", padx=20, pady=0)
-        self.querier_tab_entry_abs = ctk.CTkEntry(self, placeholder_text="Inserte texto aquí")
-        self.querier_tab_entry_abs.pack(side="top", padx=20, pady=5, fill="x")
-
-        self.querier_tab_key_lbl = ctk.CTkLabel(self, text="Keywords", anchor="w",
-                                                font=ctk.CTkFont(weight="bold"))
-        self.querier_tab_key_lbl.pack(side="top", anchor="w", padx=20, pady=0)
-        self.querier_tab_entry_key = ctk.CTkEntry(self, placeholder_text="Inserte texto aquí")
-        self.querier_tab_entry_key.pack(side="top", padx=20, pady=5, fill="x")
-
-        self.querier_tab_content_lbl = ctk.CTkLabel(self, text="Content", anchor="w",
-                                                    font=ctk.CTkFont(weight="bold"))
-        self.querier_tab_content_lbl.pack(side="top", anchor="w", padx=20, pady=0)
-        self.querier_tab_entry_content = ctk.CTkEntry(self,
-                                                      placeholder_text="Inserte texto aquí")
-        self.querier_tab_entry_content.pack(side="top", padx=20,  pady=5, fill="x")
-        # Entry boxes
-
-        self.querier_tab_search_btn = ctk.CTkButton(self, text="Search!",
-                                                    command=self.querier_tab_search_btn_event)
-        self.querier_tab_search_btn.pack(side="top", padx=20, pady=5)
-
-
-    def querier_tab_search_btn_event(self):
-        # TODO: Work In Progress
-        texto = ""
-        texto += f'Title({self.querier_tab_entry_title.get()}) - '
-        texto += f'Abs({self.querier_tab_entry_abs.get()}) - '
-        texto += f'Key({self.querier_tab_entry_key.get()}) - '
-        texto += f'Content({self.querier_tab_entry_content.get()})'
-        # HACK: If search is empty then show help. Just a demo of subproces run
-        if texto == 'Title() - Abs() - Key() - Content()':
-            excecute("python querier.py -h")
-        else:
-            print(f'{datetime.now().strftime("%Y-%m-%d %H:%M:%S")} - BUSCAAAARRR\n{texto}\n')
-        #print("BUSCAARRR!!!!")
-
-
-class IndexerPage(ctk.CTkFrame):
-    def __init__(self, parent, controller):
-        ctk.CTkFrame.__init__(self, parent)
-        self.controller = controller
-        label = ctk.CTkLabel(self, text="This is Indexer Page")
-        label.pack(side="top", fill="x", pady=10)
-        button = ctk.CTkButton(self, text="Go to the start page",
-                           command=self.dummy_action)
-        button.pack()
-
-    def dummy_action(self):
-        print("dummy function wip")
-        pass
-
-
-class HistoryPage(ctk.CTkFrame):
-    def __init__(self, parent, controller):
-        ctk.CTkFrame.__init__(self, parent)
-        self.controller = controller
-        label = ctk.CTkLabel(self, text="This is History Page")
-        label.pack(side="top", fill="x", pady=10)
-        button = ctk.CTkButton(self, text="Go to the start page",
-                           command=self.dummy_action)
-        button.pack()
-
-    def dummy_action(self):
-        print("dummy function wip")
-        pass
-
-
-class LogsFrame(ctk.CTkFrame):
-    PADX = 20
-    def __init__(self, parent, controller):
-        ctk.CTkFrame.__init__(self, parent)
-        self.controller = controller
-        self.logs_lbl = ctk.CTkLabel(self, text="Results and logs", anchor="w", font=ctk.CTkFont(weight="bold"))
-        self.logs_lbl.pack(side="top", padx=LogsFrame.PADX, pady=(10,0))
-        self.logs_box = ctk.CTkTextbox(self, width=2000, height=150, activate_scrollbars=True)
-        self.logs_box.pack(side="top", padx=LogsFrame.PADX, pady=(10, 20))
-        self.logs_box.configure(state="disabled")
-        self.show_logs_checkbox = None
-        self.stdout_redirector = StdoutRedirector(self.logs_box)
-
-    def Get_enabling_checkBox(self, parent) -> ctk.CTkCheckBox:
-        if (self.show_logs_checkbox is None):
-            self.show_logs_checkbox = ctk.CTkCheckBox(parent, text="Show results and logs", command=self._checkbox_event)
-        if bool(self.grid_info()): # This method led to know if the widget is visible now and so then syncronise the checkbox
-            self.show_logs_checkbox.select()
-        return self.show_logs_checkbox
-
-    def _checkbox_event(self):
-        value = self.show_logs_checkbox.get()
-        # Redirect the standar output between default stdout and self.logs_box
-        self.stdout_redirector.enable( value )
-        # Handle visibility on the front
-        self.grid() if value else self.grid_remove()
-        print("_checkbox_event( view_enabled={} )".format(value))
-
-
-import sys
-from typing import TextIO
-class StdoutRedirector(TextIO):
-    def __init__(self,text_widget:ctk.CTkTextbox):
-        self._text_space = text_widget
-        #self._text_space.tag_config(tagName='ERROR', background="black", foreground="red")
-        self._default_stdout = None
-        self.enable()
-
-    def write(self, string):
-            # FIXME: Next line would be preffered to keep config after writing but get next incoherent exception
-            #        ValueError: 'state' is not a supported argument. Look at the documentation for supported arguments.
-            # prev_state = self._text_space.cget("state")
-            self._text_space.configure(state="normal")
-            #try:
-            #    # FIXME: No aplica el tag. De todas formas este codigo no se lo deseo ni a mi peor enemigo, hay que mejorar esta abominacion
-            #    start_idx = string.index(bcolors.FAIL)
-            #    string = string.replace(bcolors.FAIL,"")
-            #    end_idx = string.index(bcolors.ENDC)
-            #    string = string.replace(bcolors.ENDC,"")
-            #    self._text_space.insert('end', string)
-            #    self._text_space.tag_add(tagName='ERROR', index1=int(start_idx), index2=int(end_idx))
-            #except:
-            #    pass
-            self._text_space.insert('end', string)
-            self._text_space.see('end')
-            #self._text_space.configure(prev_state)
-            self._text_space.configure(state="disabled")
-
-    def enable(self, on:bool=True):
-        if on:
-            self._default_stdout = sys.stdout
-            sys.stdout = self
-        else:
-            sys.stdout = self._default_stdout
-        self._enabled = on
-
-    def disable(self):
-        self.enable(False)
-
-
-
-class bcolors:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKCYAN = '\033[96m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
-
-def excecute(cmd) -> int:
-    #proc = subprocess.Popen( "python querier.py -h".split(), text=True, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
-    #print( proc.stdout.read() )
-    proc = subprocess.Popen( cmd.split(), text=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE )
-    out, err = proc.communicate()
-    if len(out) > 0: print( out )
-    if len(err) > 0: print( bcolors.FAIL + err + bcolors.ENDC ) # print in red
-    return proc.returncode
-
 
 
 if __name__ == "__main__":
